requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LangChain core
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# LLM Providers (install the ones you need)
langchain-openai>=0.1.0        # for OpenAI
openai>=1.30.0                 # for OpenAI
langchain-groq>=0.1.0          # for Groq
langchain-google-genai>=1.0.0  # for Gemini

# Integrations
simple-salesforce>=1.12.0
requests>=2.31.0
twilio>=9.0.0
redis>=5.0.0
